session11/requirements.txt
fastapi
uvicorn[standard]
strawberry-graphql[fastapi]
sqlalchemy
psycopg2-binary